Added BoardPanel.java
This class extends JPanel and serves as a container class for the Cell
class.

This will display state based only on Square[][] board. It has no other
information apart from the present board state. (Not even the current
player)

Move generated by the player will made made here and sends it to the
updater to verify/execute the move.

Drag and Drop if implemented will be implemented here